refactor(demo): examples except datepicker and server done
