+++ packages/website/src/SideNavigation.tsx
@@ -101,9 +101,6 @@ export const Navigation: React.FunctionComponent = () => {
                 <NavLink href="/layout/SplitLayout" label="Split layout" />
                 <NavLink href="/layout/StickyFooter" label="Sticky footer" />
                 <NavLink href="/layout/Table" label="Table" />
-                <NavLink href="/layout/TableHOC" label="TableHOC" />
-                <NavLink href="/layout/TableHOCServer" label="Table server" />
-                <NavLink href="/layout/TableHOCwithBlankSlate" label="Table blank slate" />
             </CollapsibleSideSection>
             <CollapsibleSideSection title="Form">
                 <NavLink href="/form/ActionableItem" label="Actionable item" />
